Create Lists-LearnPython
This is a solution for the exercise on http://www.learnpython.org/Lists. There was previously no solution offered. I have not checked with the others yet but will be happy to submit solutions for any! Sorry, I'm new on GitHub and honestly don't have a handle on it yet. I just wanted to offer help with Python!
-Philip