Example fix, Exercise update
I added the import to the Example code. I added an Exercise area and implemented instructions. I removed the solution from the Tutorial Code and added and solution line with the solution in it, in order to impart more critical thinking from the reader instead of just having the solution immediately.