Corrected Solution
In the original, it was numbers.length, which would return the number 3, so it was trying to access the fourth spot of the chars[] array, which doesn't exist. Changing this to numbers.length-1 would have it return 2, accessing the third spot which does exist. A simple error it seems but needs to be fixed.