Update Inheritance.md
From:
prefix the method with `super'. It is also possible to write a subclass constructor that invokes the constructor of the superclass, either implicitly or by using the `super`

To:
prefix the method with `super`. It is also possible to write a subclass constructor that invokes the constructor of the superclass, either implicitly or by using the `super`